Issue #1792 splitted up contacts to ComposerContacts
